+++ test/global-teardown.js
@@ -20,4 +20,6 @@ module.exports = async function globalTeardown() {
   killServer('default');
   killServer('hbs');
   killServer('secured');
+  killServer('prod');
+  killServer('prod-hbs');
 };